requirements.txt
bob.ap==2.1.1
bob.blitz==2.0.11
bob.core==2.1.6
bob.extension==2.3.4
bob.sp==2.0.7
numpy==1.11.2
pandoc==1.0.0b2
ply==3.9
pyssp==0.1.6.5
repoze.lru==0.6
scikits.talkbox==0.2.5
scipy==0.18.1
SpeechRecognition==3.5.0
